--- application/next-frontend/middleware1.ts
@@ -1,10 +1,0 @@
-import { NextFetchEvent, NextRequest, NextResponse } from 'next/server'
-
-type Environment = 'production' | 'development' | 'test'
-export function middleware(req: NextRequest, ev: NextFetchEvent) {
-    const currentEnv: Environment = process.env.NODE_ENV
-    if (currentEnv === 'production' && req.headers.get('x-forwarded-proto') !== 'https') {
-        return NextResponse.redirect(`https://${req.headers.get('host')}${req.nextUrl.pathname}`, 301)
-    }
-    return NextResponse.next()
-}
